[ ]RTKCRM-7.2 WEB-морда - jsp - IN PROGRESS

+++ CalculatorMvc/src/main/java/ru/nvd/andr/calcmvc/service/CalcOperationsService.java
@@ -9,6 +9,6 @@ public interface CalcOperationsService {
 
     public List<TableCalcOperations> listTableCalcOperations();
 
-    public void removeTableCalcOperations(Integer objid);
+    public void removeTableCalcOperations(Long objid);
 
 }
